chore(sdk-coin-near): move account-lib code
- moves account-lib code to sdk-coin-near module

Ticket: BG-48154
