+++ modules/account-lib/tsconfig.json
@@ -32,6 +32,9 @@
       "path": "../sdk-coin-eth"
     },
     {
+      "path": "../sdk-coin-near"
+    },
+    {
       "path": "../sdk-coin-rbtc"
     },
     {
