+++ modules/sdk-coin-near/package.json
@@ -41,7 +41,15 @@
     ]
   },
   "dependencies": {
-    "@bitgo/sdk-core": "^1.0.1"
+    "@bitgo/sdk-core": "^1.0.1",
+    "@bitgo/statics": "^6.17.0",
+    "@stablelib/hex": "^1.0.0",
+    "bignumber.js": "^9.0.0",
+    "bn.js": "^5.2.1",
+    "bs58": "^4.0.1",
+    "js-sha256": "^0.9.0",
+    "near-api-js": "^0.44.2",
+    "tweetnacl": "^1.0.3"
   },
   "devDependencies": {
     "@bitgo/sdk-api": "^1.0.1",
