+++ modules/sdk-coin-near/src/index.ts
@@ -1,2 +1,1 @@
-export * from './near';
-export * from './tnear';
+export * from './lib';
