+++ modules/sdk-coin-near/src/lib/constants.ts
