+++ modules/sdk-coin-near/src/lib/errors.ts
