+++ modules/sdk-coin-near/src/lib/iface.ts
