+++ modules/sdk-coin-near/src/lib/index.ts
