+++ modules/sdk-coin-near/src/lib/stakingActivateBuilder.ts
