+++ modules/sdk-coin-near/src/lib/walletInitializationBuilder.ts
