+++ modules/sdk-coin-near/tsconfig.json
@@ -18,6 +18,9 @@
     },
     {
       "path": "../sdk-test"
+    },
+    {
+      "path": "../statics"
     }
   ]
 }
